ajustando controllers com as responses corretas

+++ src/main/java/tcc/sgmeabackend/service/AbstractService.java
@@ -2,8 +2,8 @@ package tcc.sgmeabackend.service;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import tcc.sgmeabackend.service.exceptions.ResourceNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,7 +11,6 @@ import java.util.Optional;
 public abstract class AbstractService<T> implements Rest<T> {
 
     protected abstract JpaRepository<T, String> getRepository();
-
 
     @Override
     public List<T> findAll() {
@@ -34,16 +33,21 @@ public abstract class AbstractService<T> implements Rest<T> {
 
     @Override
     public T update(String id, T resource) {
-        return this.getRepository().save(resource);
+        Optional<T> existingResource = this.getRepository().findById(id);
+        if (existingResource.isPresent()) {
+            return this.getRepository().save(resource);
+        } else {
+            throw new ResourceNotFoundException("Resource with id " + id + " not found");
+        }
     }
 
     @Override
     public void delete(String id) {
-        this.getRepository().deleteById(id);
-    }
-
-    @Override
-    public Page<T> findAll(Pageable pageable) {
-        return null;
+        Optional<T> existingResource = this.getRepository().findById(id);
+        if (existingResource.isPresent()) {
+            this.getRepository().deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Resource with id " + id + " not found");
+        }
     }
 }
